+++ foxtive/src/helpers/mod.rs
@@ -103,6 +103,7 @@ mod tokio;
 
 #[cfg(feature = "regex")]
 mod regex;
+pub mod env;
 
 pub use tokio::blk;