+++ foxtive/src/rabbitmq/conn.rs
@@ -1,11 +1,13 @@
-use std::env;
-
+use crate::prelude::AppResult;
+use anyhow::Error;
 use deadpool_lapin::{Manager, Pool};
 use lapin::ConnectionProperties;
 
-pub async fn establish_rabbit_connection_pool(env_prefix: &String) -> Pool {
-    let dsn: String = env::var(format!("{}_RMQ_DSN", env_prefix)).unwrap();
-
+pub async fn create_rmq_conn_pool(dsn: &str, pool_max_size: usize) -> AppResult<Pool> {
     let manager = Manager::new(dsn, ConnectionProperties::default());
-    Pool::builder(manager).max_size(10).build().unwrap()
+
+    Pool::builder(manager)
+        .max_size(pool_max_size)
+        .build()
+        .map_err(Error::msg)
 }